Refactoring for jenkins client (#366)
